Separated Authentication Routes into authRoutes.js file inside auth_helpers
